+++ app.blocks/handler/handler.deps.js
@@ -5,6 +5,7 @@
     ],
     shouldDeps : [
         { block : 'util' },
+        { block : 'request' },
         { block : 'view' },
         { block : 'http', elem : 'error' }
     ]