+++ examples/silly/common.blocks/page/page.node.js
@@ -11,7 +11,7 @@ App.View.decl('page', {
      * @returns {String|Buffer}
      */
     render : function(ctx) {
-        App.Logger.log('å page handler is runnig');
+        App.Logger.debug('å page handler is runnig');
 
         var req = ctx.request;