+++ HISTORY.md
@@ -1,3 +1,15 @@
+
+0.1.3, 2013-02-24
+-----------------
+
+  - "request" block added for request normalization API
+  - "url" block removed (use "request")
+  - "state" block removed (use "request")
+  - App.Http.start() method was renamed to App.Http.run()
+  - App.Logger.log() method was renamed to App.Logger.debug()
+  - some inaccuracies were fixed in README.md
+  - code refactoring
+
 0.1.1, 2013-02-18
 -----------------